+++ lib/controller/controller.py
@@ -374,7 +374,9 @@ class Controller:
                 directory_path = self.setup_batch_reports()
                 filename = "BATCH." + self.get_output_extension()
             else:
-                parsed = urlparse(options["urls"][0])
+                self.set_target(options["urls"][0])
+
+                parsed = urlparse(self.url)
 
                 if not parsed.netloc:
                     parsed = urlparse(f"//{options['urls'][0]}")
@@ -565,12 +567,15 @@ class Controller:
             except KeyboardInterrupt:
                 self.handle_pause()
 
+            time.sleep(0.3)
+
     def add_directory(self, path):
         """Add directory to the recursion queue"""
 
         # Pass if path is in exclusive directories
         if any(
-            "/" + dir in path for dir in options["exclude_subdirs"]
+            path.startswith(dir) or "/" + dir in path
+            for dir in options["exclude_subdirs"]
         ):
             return